Implemented a runtime function to get CUDA SM count.
In the future, this can be extended to get information about the layout
of other devices as well. Also, the number of total threads could be a
usefull insight, so a template for that has been added.

The return values for HSA and Level Zero return core counts for 7900 XTX
and Arc A770 respectively.